Fix import cycle issue
